+++ src/policy/handcrafted/spell_profile.rs
@@ -473,6 +473,7 @@ impl HandcraftedPolicy {
             | EffectDef::GainClassLevel { .. }
             | EffectDef::SubstituteBasicLandTypeUntilEndOfTurn { .. }
             | EffectDef::CreateEmblem { .. }
+            | EffectDef::CreateOngoingEffect(_)
             | EffectDef::ReturnWithHasteAndFinality { .. }
             | EffectDef::Transform { .. }
             | EffectDef::ScheduleTurnPhases(_)